Rename VesselReportings to CardReportingList
