+++ frontend/src/features/Reporting/components/CardReportingList/Current/ConfirmDeletionModal.tsx
