+++ frontend/src/features/Reporting/components/CardReportingList/Summary/InfractionSuspicionSummary.tsx
